Tidy up some of the changes
