+++ include/Makefile.am
@@ -1,17 +1,56 @@
 
 cybermonincdir = ${includedir}/cybermon
 
-cybermoninc_HEADERS = cybermon/address.h cybermon/base_context.h \
-	cybermon/ber.h cybermon/context.h cybermon/cybermon-lua.h \
-	cybermon/dns.h cybermon/dns_protocol.h cybermon/engine.h \
-	cybermon/etsi_li.h cybermon/exception.h cybermon/flow.h \
-	cybermon/forgery.h cybermon/ftp.h cybermon/hexdump.h \
-	cybermon/http.h cybermon/icmp.h cybermon/ip.h \
-	cybermon/manager.h cybermon/monitor.h cybermon/nhis11.h \
-	cybermon/ntp.h cybermon/ntp_protocol.h \
-	cybermon/observer.h cybermon/packet_capture.h \
-	cybermon/packet.h cybermon/pdu.h cybermon/protocol.h \
-	cybermon/reaper.h cybermon/resource.h cybermon/smtp.h \
-	cybermon/socket.h cybermon/specification.h \
-	cybermon/tcp.h cybermon/thread.h cybermon/transport.h cybermon/udp.h \
-	cybermon/unrecognised.h cybermon/xml.h cybermon/serial.h
+cybermoninc_HEADERS = cybermon/address.h
+	cybermon/base_context.h \
+	cybermon/ber.h \
+	cybermon/context.h \
+	cybermon/cybermon-lua.h \
+	cybermon/dns_context.h \
+	cybermon/dns_protocol.h  \
+	cybermon/dns_over_tcp.h \
+	cybermon/dns_over_udp.h \
+	cybermon/engine.h \
+	cybermon/etsi_li.h \
+	cybermon/exception.h \
+	cybermon/flow.h \
+	cybermon/forgery.h \
+	cybermon/ftp.h \
+	cybermon/hexdump.h \
+	cybermon/http.h \
+	cybermon/icmp.h \
+	cybermon/imap.h \
+	cybermon/imap_context.h \
+	cybermon/imap_ssl.h \
+	cybermon/imap_ssl_context.h \
+	cybermon/ip.h \
+	cybermon/manager.h \
+	cybermon/monitor.h \
+	cybermon/nhis11.h \
+	cybermon/ntp.h \
+	cybermon/ntp_protocol.h \
+	cybermon/observer.h \
+	cybermon/packet_capture.h \
+	cybermon/packet.h \
+	cybermon/pdu.h \
+	cybermon/pop3.h \
+	cybermon/pop3_context.h \
+	cybermon/pop3_ssl.h \
+	cybermon/pop3_ssl_context.h \
+	cybermon/protocol.h \
+	cybermon/reaper.h \
+	cybermon/resource.h \
+	cybermon/serial.h \
+	cybermon/smtp.h \
+	cybermon/smtp_auth.h \
+	cybermon/smtp_auth_context.h
+	cybermon/socket.h \
+	cybermon/specification.h \
+	cybermon/tcp.h \
+	cybermon/tcp_ports.h \
+	cybermon/thread.h \
+	cybermon/transport.h \
+	cybermon/udp.h \
+	cybermon/udp_ports.h \
+	cybermon/unrecognised.h \
+	cybermon/xml.h
