+++ include/cybermon/cybermon-lua.h
@@ -502,19 +502,19 @@ namespace cybermon {
 			  int status,
 			  const std::list<std::string>& responses);
 
-   void dns_over_tcp_message(engine& an, const context_ptr cf,
-                             const dns_header& hdr,
-                             const std::list<dns_query> queries,
-                             const std::list<dns_rr> answers,
-                             const std::list<dns_rr> authorities,
-                             const std::list<dns_rr> additional);
+    void dns_over_tcp_message(engine& an, const context_ptr cf,
+                                const dns_header& hdr,
+                                const std::list<dns_query> queries,
+                                const std::list<dns_rr> answers,
+                                const std::list<dns_rr> authorities,
+                                const std::list<dns_rr> additional);
 
 	void dns_over_udp_message(engine& an, const context_ptr cf,
-                            const dns_header& hdr, 
-                            const std::list<dns_query> queries,
-                            const std::list<dns_rr> answers,
-                            const std::list<dns_rr> authorities,
-                            const std::list<dns_rr> additional);
+                                const dns_header& hdr, 
+                                const std::list<dns_query> queries,
+                                const std::list<dns_rr> answers,
+                                const std::list<dns_rr> authorities,
+                                const std::list<dns_rr> additional);
 
     void ntp_timestamp_message(engine& an, const context_ptr cf,
 			                   const ntp_timestamp& ts);
