+++ include/cybermon/dns_over_tcp.h
@@ -8,15 +8,11 @@
 #ifndef CYBERMON_DNS_OVER_TCP_H
 #define CYBERMON_DNS_OVER_TCP_H
 
-#include <stdint.h>
-
-#include <set>
 
 #include "context.h"
+#include "pdu.h"
 #include "manager.h"
-#include "serial.h"
-#include "protocol.h"
-#include "dns_protocol.h"
+
 
 namespace cybermon
 {
