+++ src/smtp_auth.C
@@ -4,9 +4,7 @@
 #include <cybermon/address.h>
 #include <cybermon/flow.h>
 #include <cybermon/manager.h>
-#include <cybermon/unrecognised.h>
-#include <cybermon/dns.h>
-#include <cybermon/ntp.h>
+#include <cybermon/smtp_auth_context.h>
 
 
 using namespace cybermon;
