+++ src/udp.C
@@ -1,12 +1,11 @@
 
 #include <cybermon/udp.h>
+
 #include <cybermon/address.h>
 #include <cybermon/flow.h>
 #include <cybermon/manager.h>
-#include <cybermon/unrecognised.h>
-#include <cybermon/dns.h>
-#include <cybermon/ntp.h>
 #include <cybermon/udp_ports.h>
+#include <cybermon/unrecognised.h>
 
 
 using namespace cybermon;
